refactor(react): update Accordion to use config
