+++ packages/react/src/index.ts
@@ -6,6 +6,7 @@
 // components
 
 export * from './components/Accordion'
+export * from './components/Accordion.client'
 export * from './components/AccordionItemGroup'
 export * from './components/Admonition'
 export * from './components/Avatar'
