+++ tests/react/components/accordion.test.tsx
@@ -6,24 +6,29 @@ import {
   AccordionItemTrigger,
   AccordionItemIndicator,
   AccordionItemContent,
+  CerberusProvider,
 } from '@cerberus-design/react'
-import { setupStrictMode } from '@/utils'
+import { makeConfig, setupStrictMode } from '@/utils'
 
 describe('Accordion', () => {
   setupStrictMode()
   afterEach(cleanup)
 
+  const config = makeConfig()
+
   test('should render an accordion element', () => {
     render(
-      <Accordion>
-        <AccordionItem value="one">
-          <AccordionItemTrigger>
-            Accordion Item 1
-            <AccordionItemIndicator />
-          </AccordionItemTrigger>
-          <AccordionItemContent>Accordion content</AccordionItemContent>
-        </AccordionItem>
-      </Accordion>,
+      <CerberusProvider config={config}>
+        <Accordion>
+          <AccordionItem value="one">
+            <AccordionItemTrigger>
+              Accordion Item 1
+              <AccordionItemIndicator />
+            </AccordionItemTrigger>
+            <AccordionItemContent>Accordion content</AccordionItemContent>
+          </AccordionItem>
+        </Accordion>
+      </CerberusProvider>,
     )
     expect(screen.getByText(/Accordion Item 1/i)).toBeTruthy()
     expect(screen.getByText(/Accordion content/i)).toBeTruthy()
